requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
autopep8==1.6.0
billiard==4.2.4
boto3==1.42.27
botocore==1.42.27
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
celery==5.6.0
certifi==2025.4.26
cffi==2.0.0
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.4
cryptography==46.0.3
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2
django-allauth==65.11.2
django-cors-headers==3.11.0
django-crispy-forms==2.4
django-environ==0.12.0
django-humanize==0.1.2
django-jazzmin==3.0.1
django-sendgrid-v5==1.3.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
firebase-admin==6.8.0
fonttools==4.61.1
google-api-core==2.25.0
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
gunicorn==23.0.0
hiredis==3.3.0
html5lib==1.1
httplib2==0.22.0
humanize==4.12.3
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
joblib==1.5.2
kombu==5.6.1
lxml==5.4.0
MarkupSafe==3.0.3
msgpack==1.1.0
numpy==2.3.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.31.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pydyf==0.12.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.9.0
PyMuPDF==1.26.4
pyparsing==3.2.3
pypdf==5.6.0
pyphen==0.17.2
pytesseract==0.3.13
pytest==8.3.5
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.1
requests==2.32.3
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.7.1
scipy==1.16.1
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tesseract==0.1.3
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
toml==0.10.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.4.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.14
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.5
wfastcgi==3.0.0
whitenoise==6.11.0
xhtml2pdf==0.2.17
xlsxwriter==3.2.9
zopfli==0.4.0
